Camera + New Movement
This is a new camera system and movement because I was having a hard time getting the player to rotate with the camera so now the player should work properly. IMPORTANT !!!! When making ground label it as Ground in the inspector for the Layer section of the ground object. The jump has also been changed as using the rigidbody along with the character controller was not functioning.