Enemy2prefab/PlayerContainerChange
I added a new enemy that only allow the player to follow the player if the player isn't looking at it. also the other change that I made in the player container is putting the arm as a child of the main camera so the arm is moving as well.